Lessons_AzureOpenAI/requirements.txt
ipykernel
pandas
numpy
streamlit
openai
python-dotenv
requests
uuid
faker
fastapi
uvicorn
mlflow
PyPDF2
sqlalchemy
llama-index
llama-index-embeddings-azure-openai
IPython
faker
openpyxl
xlrd
docling